feat: add tests for head
